+++ src/app/home/aboutMe.tsx
@@ -1,127 +1,78 @@
-"use client"
-
-import { css } from '@linaria/core';
-import { styled } from '@linaria/react';
-import { useState } from 'react';
-import { ImInfo } from "react-icons/im";
-import { MdLibraryMusic, MdConstruction } from "react-icons/md";
-import { BsEnvelopePaperFill } from "react-icons/bs";
-
-import { scaleDown, scaleUp } from '@/components/animations';
-import Socials from '@/components/socials';
-import Name from './name';
-import DraggableWindow from '@/components/DraggableWindow';
-
-interface props {
-  name: string;
-  Icon: any;
-}
+import { styled } from "@linaria/react"
+import { scaleUp } from "@/components/animations"
+import { css } from "@linaria/core"
+import EpicCarlito from "../../../public/epiccarlito.svg";
+import Face from "../../../public/face.png";
+import Image from "next/image";
 
 const Container = styled.div`
-  flex-grow: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const ItemContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  background-color: #1f2937;
-  border: 1px solid var(--white-but-not);
-  padding: 2rem;
-  gap: 0.75rem;
-
-  @media (min-width: 48rem) {
-    padding: 3rem;
-  }
+    display: flex;
+    flex-direction: column;
 `
 
-const Buttons = styled.div`
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-
-  @media (min-width: 48rem) {
-    grid-template-columns: repeat(4, minmax(0, 1fr));
-  }
-`
-
-const ButtonContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.25rem;
-`
-
-const ButtonIcon = css`
-  cursor: pointer;
-  height: 3rem;
-  width: 3rem;
-
-  @media (min-width: 48rem) {
-    height: 4rem;
-    width: 4rem;
-  }
-`
-
-const ButtonText = styled.p`
-  font-size: var(--text-sm);
-  font-weight: bolder;
-`
-
-const SocialsDivider = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-block: 0.25rem;
+const TopSection = styled.div`
+    display: flex;
+    align-items: center;
+    margin: 1rem 1.5rem;
 `
 
 const Divider = styled.div`
-  flex-grow: 1;
-  border-top: 2px solid var(--white-but-not);
+    border-bottom: 1px solid var(--white-but-not);
 `
 
-const SocialsText = styled.span`
-  font-size: var(--text-base);
-  padding-inline: 0.75rem;
+const BottomSection = styled.div`
+    margin: 1rem 1.5rem;
+
 `
 
-export default function Page() {
-  return (
-    <>
-      <Container>
-        <ItemContainer>
-          <Name />
-          <Buttons>
-            <ToggleButton name='about' Icon={ImInfo} />
-            <ToggleButton name='música' Icon={MdLibraryMusic} />
-            <ToggleButton name='projects' Icon={MdConstruction} />
-            <ToggleButton name='poems' Icon={BsEnvelopePaperFill} />
-          </Buttons>
-          <SocialsDivider>
-            <Divider />
-            <SocialsText>Socials</SocialsText>
-            <Divider />
-          </SocialsDivider>
-          <Socials />
-        </ItemContainer>
-      </Container>
-    </>
-  );
-}
+const ProfilePictures = styled.div`
+  display: flex;
+  flex-direction: row;
+`
 
-function ToggleButton(props: props) {
-  const { name, Icon } = props;
-  const [showPopup, setShowPopup] = useState(false);
+const ProfilePicture = css`
+  height: 6rem;
+  width: 6rem;
+  border: 1px solid black;
+  border-radius: 9999px;
+  filter: drop-shadow(3px 3px 5px rgb(0 0 0 / 0.50));
+  z-index: 2;
 
-  return (
-    <>
-      <ButtonContainer onClick={() => setShowPopup(!showPopup)}>
-        <Icon className={`${ButtonIcon} ${scaleUp} ${scaleDown}`} />
-        <ButtonText>{name}</ButtonText>
-      </ButtonContainer>
+  @media (min-width: 48rem) {
+    height: 8rem;
+    width: 8rem;
+  }
+`
 
-      {showPopup && (<DraggableWindow name={name} setState={setShowPopup} />)}
-    </>
-  )
+const TextSection = styled.div`
+    display: flex;
+    flex-direction: column;
+    padding-left: 1rem;
+`
+
+export default function AboutMe() {
+    return (
+        <Container>
+            <TopSection>
+                <ProfilePictures>
+                    <EpicCarlito
+                        className={`${ProfilePicture} ${scaleUp}`} />
+                    <Image
+                        src={Face}
+                        alt="IRL FACE WHAT?!?!?"
+                        className={`${ProfilePicture} ${scaleUp}`}
+                        style={{ marginLeft: "-0.5rem" }}
+                    />
+                </ProfilePictures>
+                <TextSection>
+                <h1>Hola! I am James!</h1>
+                <p>Figuring out my code along the way ¯\_(ツ)_/¯</p>
+                </TextSection>
+            </TopSection>
+            <Divider/>
+            <BottomSection>
+                <p>Help me out here</p>
+            </BottomSection>
+        </Container>
+    )
 }
