+++ src/app/home/components/animatedWaves.tsx
