+++ src/components/animations.tsx
@@ -1,7 +1,7 @@
 import { css } from "@linaria/core";
 
 const scaleUp = css`
-  transition: transform 0.3s ease-in-out;
+  transition: transform 0.2s ease-in-out;
 
   &:hover {
     transform: scale(1.1);
@@ -9,7 +9,7 @@ const scaleUp = css`
 `;
 
 const scaleDown = css`
-  transition: transform 0.3s ease-in-out;
+  transition: transform 0.2s ease-in-out;
 
   &:active {
     transform: scale(0.9);
